course/Task5/requirements.txt
streamlit
pandas
numpy
scipy
plotly
fpdf2
openpyxl
kaleido
